# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "NFT Music",
            "url": "https://nft.threenext.com",
            "email": "seki@threenext.com"
        }
    },
    "paths": {
...
        "/users": {
            "get": {
                "description": "NFTミュージックのアカウントを登録する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "アカウント"
                ],
                "summary": "NFTミュージックのアカウントを登録する",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ports.UserOutput"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "post": {
                "description": "NFTミュージックのアカウントを登録する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "アカウント"
                ],
                "summary": "NFTミュージックのアカウントを登録する",
                "parameters": [
                    {
                        "description": "ユーザー情報",
                        "name": "users",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.UserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.UserOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
        "/users/wallet/{wallet}": {
            "get": {
                "description": "ウォレットアドレスをキーにNFTミュージックのアカウント情報を取得する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "アカウント"
                ],
                "summary": "ウォレットアドレスからアカウント情報を取得する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ウォレットアドレス",
                        "name": "wallet",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.UserOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
        "/users/{id}": {
            "get": {
                "description": "NFTミュージックのアカウントを登録する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "アカウント"
                ],
                "summary": "NFTミュージックのアカウントを登録する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.UserOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "put": {
                "description": "NFTミュージックのアカウントの情報を変更する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "アカウント"
                ],
                "summary": "NFTミュージックのアカウントの情報を変更する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ユーザー情報",
                        "name": "users",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ports.UserInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.UserOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            },
            "delete": {
                "description": "NFTミュージックのアカウントを削除する",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "アカウント"
                ],
                "summary": "NFTミュージックのアカウントを削除する",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ユーザーID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ports.UserOutput"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ports.ErrorResponseObject"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ports.ErrorResponseObject": {
            "type": "object",
            "properties": {
                "error_type": {
                    "type": "string",
                    "example": "リクエストの構文が不正やページが見つからない等のタイプ別"
                },
                "message": {
                    "type": "string",
                    "example": "原因となるのエラーメッセージ"
                },
                "status_code": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
...
        "ports.UserInput": {
            "type": "object",
            "required": [
                "email",
                "name",
                "wallet"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "東京都千代田区１−１−１"
                },
                "business_id": {
                    "type": "string",
                    "example": "019504de-f120-76af-905b-2fe06352b04a"
                },
                "email": {
                    "type": "string",
                    "example": "takeshi-kitagawa@gmail.com"
                },
                "eyecatch": {
                    "type": "string",
                    "example": "https://www.yahoo.com/img/test.jpg"
                },
                "face_image": {
                    "type": "string",
                    "example": "https://www.yahoo.com/img/test.jpg"
                },
                "name": {
                    "type": "string",
                    "example": "北川武"
                },
                "profile": {
                    "type": "string",
                    "maxLength": 4000,
                    "minLength": 10,
                    "example": "私はいつでも明るいです"
                },
                "role": {
                    "type": "string",
                    "example": "member"
                },
                "wallet": {
                    "type": "string",
                    "example": "0x411F8Cd035DB4874a8F1247Cb49908c794A0AB50"
                },
                "website": {
                    "type": "string",
                    "example": "https://www.yahoo.com"
                }
            }
        },
        "ports.UserOutput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "東京都千代田区１−１−１"
                },
                "business_id": {
                    "type": "string",
                    "example": "019504de-f120-76af-905b-2fe06352b04a"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-11-04T20:51:26Z"
                },
                "email": {
                    "type": "string",
                    "example": "takeshi-kitagawa@gmail.com"
                },
                "eyecatch": {
                    "type": "string",
                    "example": "https://www.yahoo.com/img/test.jpg"
                },
                "face_image": {
                    "type": "string",
                    "example": "https://www.yahoo.com/img/test.jpg"
                },
                "id": {
                    "type": "string",
                    "example": "01932563-f671-71ff-9a0d-c452de9d06aa"
                },
                "name": {
                    "type": "string",
                    "example": "北川武"
                },
                "profile": {
                    "type": "string",
                    "example": "私はいつでも明るいです"
                },
                "role": {
                    "type": "string",
                    "example": "member"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-11-04T20:51:26Z"
                },
                "wallet": {
                    "type": "string",
                    "example": "01932564-3432-7b56-b772-7c69586e1897"
                },
                "website": {
                    "type": "string",
                    "example": "https://www.yahoo.com"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        }
    }
}